+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Add `TimingFunctionType` attribute to `Animatable` to specifies the speed curve 
 
 #### Bugfixes
 
-N/A
+- Padding UIView for `PaddingDesignable` can't expand beyond `UITextField` and now has a default height of 1pt. [#483](https://github.com/IBAnimatable/IBAnimatable/pull/483) by [@SD10](https://github.com/SD10)
 
 ### [4.1.0](https://github.com/IBAnimatable/IBAnimatable/releases/tag/4.1.0)